# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quartermaster API",
    "description": "Self-hosted kitchen inventory management.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/households": {
      "post": {
        "tags": [
          "households"
        ],
        "operationId": "household_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/households/current": {
      "get": {
        "tags": [
          "households"
        ],
        "operationId": "household_current_get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "households"
        ],
        "operationId": "household_current_update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/households/current/invites": {
      "get": {
        "tags": [
          "households"
        ],
        "operationId": "household_invites_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InviteDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "households"
        ],
        "operationId": "household_invite_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/households/current/members": {
      "get": {
        "tags": [
          "households"
        ],
        "operationId": "household_members_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/households/current/members/{user_id}": {
      "delete": {
        "tags": [
          "households"
        ],
        "operationId": "household_member_remove",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invites/redeem": {
      "post": {
        "tags": [
          "households"
        ],
        "operationId": "invite_redeem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invites/{id}": {
      "delete": {
        "tags": [
          "households"
        ],
        "operationId": "invite_revoke",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable error code, machine-readable."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description. Not localised; clients should prefer `code`."
          },
          "unrestorable_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Present on `code = batch_not_restorable` when the failure came from a\nbulk call — lists the specific batches the caller needs to look at.\nAbsent on single-batch restore (the path already names the batch)."
          }
        }
      },
...
      "CreateHouseholdRequest": {
        "type": "object",
        "required": [
          "name",
          "timezone"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        }
      },
      "CreateInviteRequest": {
        "type": "object",
        "required": [
          "expires_at",
          "max_uses",
          "role_granted"
        ],
        "properties": {
          "expires_at": {
            "type": "string"
          },
          "max_uses": {
            "type": "integer",
            "format": "int64"
          },
          "role_granted": {
            "$ref": "#/components/schemas/MembershipRole"
          }
        }
      },
...
      "HouseholdDetailDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "timezone"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        }
      },
      "HouseholdDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "timezone"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        }
      },
      "InviteDto": {
        "type": "object",
        "required": [
          "id",
          "code",
          "role_granted",
          "expires_at",
          "max_uses",
          "use_count",
          "created_at"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "max_uses": {
            "type": "integer",
            "format": "int64"
          },
          "role_granted": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "use_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "MeHouseholdDto": {
        "type": "object",
        "required": [
          "household",
          "role",
          "joined_at"
        ],
        "properties": {
          "household": {
            "$ref": "#/components/schemas/HouseholdDto"
          },
          "joined_at": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "required": [
          "user",
          "households"
        ],
        "properties": {
          "household_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "household_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "household_timezone": {
            "type": [
              "string",
              "null"
            ]
          },
          "households": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeHouseholdDto"
            }
          },
          "public_base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "MemberDto": {
        "type": "object",
        "required": [
          "user",
          "role",
          "joined_at"
        ],
        "properties": {
          "joined_at": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "MembershipRole": {
        "type": "string",
        "enum": [
          "admin",
          "member"
        ]
      },
...
      "RedeemInviteRequest": {
        "type": "object",
        "required": [
          "invite_code"
        ],
        "properties": {
          "invite_code": {
            "type": "string"
          }
        }
      },
...
      "UpdateHouseholdRequest": {
        "type": "object",
        "required": [
          "name",
          "timezone"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    }
  },
...
}